+++ panorama_openedx_backend/utils.py
@@ -60,7 +60,7 @@ def get_user_arn(user: User) -> str:
     """
     user_access_configuration = UserAccessConfiguration.objects.get(user=user)
 
-    return user_access_configuration.arn or getattr(settings, "PANORAMA_DEFAULT_USER_ARN")
+    return user_access_configuration.arn or settings["PANORAMA_DEFAULT_USER_ARN"]
 
 
 def get_user_dashboards(user: User) -> list: